sync/requirements.txt
pymongo
python-dotenv
pandas
numpy
